+++ pcs/lib/commands/test/resource/test_resource_create.py
@@ -86,6 +86,29 @@ fixture_cib_resources_xml_master_simplest = """<resources>
 
 fixture_cib_resources_xml_master_simplest_disabled = """<resources>
     <master id="A-master">
+        <meta_attributes id="A-master-meta_attributes">
+            <nvpair id="A-master-meta_attributes-target-role" name="target-role"
+                value="Stopped"
+            />
+        </meta_attributes>
+        <primitive class="ocf" id="A" provider="heartbeat" type="Dummy">
+            <operations>
+                <op id="A-monitor-interval-10" interval="10" name="monitor"
+                    timeout="20"
+                />
+                <op id="A-start-interval-0s" interval="0s" name="start"
+                    timeout="20"
+                />
+                <op id="A-stop-interval-0s" interval="0s" name="stop"
+                    timeout="20"
+                />
+            </operations>
+        </primitive>
+    </master>
+</resources>"""
+
+fixture_cib_resources_xml_master_simplest_disabled_meta_after = """<resources>
+    <master id="A-master">
         <primitive class="ocf" id="A" provider="heartbeat" type="Dummy">
             <operations>
                 <op id="A-monitor-interval-10" interval="10" name="monitor"
@@ -170,6 +193,12 @@ fixture_cib_resources_xml_clone_simplest = """<resources>
 
 fixture_cib_resources_xml_clone_simplest_disabled = """<resources>
     <clone id="A-clone">
+        <meta_attributes id="A-clone-meta_attributes">
+            <nvpair id="A-clone-meta_attributes-target-role"
+                name="target-role"
+                value="Stopped"
+            />
+        </meta_attributes>
         <primitive class="ocf" id="A" provider="heartbeat" type="Dummy">
             <operations>
                 <op id="A-monitor-interval-10" interval="10" name="monitor"
@@ -183,12 +212,6 @@ fixture_cib_resources_xml_clone_simplest_disabled = """<resources>
                 />
             </operations>
         </primitive>
-        <meta_attributes id="A-clone-meta_attributes">
-            <nvpair id="A-clone-meta_attributes-target-role"
-                name="target-role"
-                value="Stopped"
-            />
-        </meta_attributes>
     </clone>
 </resources>"""
 
@@ -606,7 +629,7 @@ class CreateAsMaster(CommonResourceTest):
                 wait="10",
                 master_meta_options={"target-role": "Stopped"}
             ),
-            fixture_cib_resources_xml_master_simplest_disabled,
+            fixture_cib_resources_xml_master_simplest_disabled_meta_after,
             fixture_state_resources_xml(role="Stopped"),
         )